requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.2.2
djangorestframework==3.14.0
gunicorn==20.1.0
mssql-django==1.3
psycopg2-binary==2.9.6
pyodbc==4.0.39
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
static3==0.7.0
typing_extensions==4.6.3
